refactor: factored remaining Tag methods into BaseTag, changed element event object sentinels into a neater enum
